refactor: Add UserRole enum to User model
This code change adds a new enum class `UserRole` to the `User` model in the `models.py` file. The `UserRole` enum defines two roles: `admin` and `user`. The `role` field in the `User` model is updated to use the `UserRole` enum with a default value of `user`.

This commit message follows the established convention of using a prefix (`refactor`) to indicate a code refactoring or improvement.