refactor: Add user role to signup_user function
This code change updates the `signup_user` function in the `curd.py` file of the User Service. It adds the `role` parameter to the function signature and assigns it to the `role` field of the `new_user` object. This allows the `role` to be specified when signing up a new user.

This commit message follows the established convention of using a prefix (`refactor`) to indicate a code refactoring or improvement.