refactor: Added Notification Service
This code change updates the Dockerfile and pyproject.toml for the Notification Service. In the Dockerfile, the `poetry config virtualenvs.create false` command is modified to include the `--local` flag. This ensures that the virtual environment is not created locally. In the pyproject.toml file, new dependencies are added for FastAPI, SQLModel, Uvicorn, Passlib, Psycopg2, Python-Jose, Requests, FastAPI-Mail, and Email-Validator. These changes improve the functionality and performance of the Notification Service.

Commit message follows the established convention of using a prefix (`refactor`) to indicate a code refactoring or improvement.