refactor: Update Dockerfile and pyproject.toml for Notification Service
This code change modifies the Dockerfile and pyproject.toml for the Notification Service. In the Dockerfile, the `poetry config virtualenvs.create false` command is updated to remove the `--local` flag. This change ensures that the virtual environment is not created locally. In the pyproject.toml file, new dependencies are added for FastAPI, SQLModel, Uvicorn, Passlib, Psycopg2, Python-Jose, Requests, FastAPI-Mail, and Email-Validator. These updates improve the functionality and performance of the Notification Service.