refactor: add email notfication service
This code change updates the dependencies for the Notification Service in the pyproject.toml file. The aiokafka and protobuf packages are added with their respective versions. These updates enhance the functionality and performance of the Notification Service.

Commit message follows the established convention of using a prefix (`refactor`) to indicate a code refactoring or improvement.