refactor: Add Kafka producer for service3
This code change adds a Kafka producer to the service3 module. The producer is responsible for producing messages to a Kafka topic using the AIOKafkaProducer library. The producer is started and stopped using the `produce_message` function. This addition enables service3 to send messages to other services or consumers through Kafka.

Commit message follows the established convention of using a prefix (`refactor`) to indicate a code refactoring or improvement.