refactor: Update Dapr Kafka Component for user-service
This code change updates the deployment workflow to deploy the Dapr Kafka Component for the user-service. The `az containerapp env dapr-component set` command has been modified to include the `--dapr-component-name` flag with the value `user-service` and the `--yaml` flag with the value `./user-service/components/pubsub.yaml`. This ensures that the Dapr Kafka Component is correctly deployed for the user-service.

Commit message follows the established convention of using a prefix (`refactor`) to indicate a code refactoring or improvement.